Add links & capitalization
These changes are intended to make the Readme easier to understand.

I have emphasized the difference between PyPy and CPython.

Adding links will also reduce confusion if a user does not know what e.g. PyPy and Mamba are.
Mamba in particular is difficult to search for, because it shares a name with at least 2 other
established Python-related projects.