Use temporary default GITHUB_TOKEN for update PRs
By extending the scope of the per run temporaray default GitHub action
token `GITHUB_OTKEN` the need of using a ssh key for cloning/pushing updates
is eliminated.

So no need to manage a ssh anymore and this is also a security improvement
as the ssh key could be leaked, while the token is only valid during the
workflow run.

This got possible via the changes announced here a while ago: https://github.blog/changelog/2021-04-20-github-actions-control-permissions-for-github_token/